#127 - Add: update taxonomy and Update: taxonomy table
